refactor: beneficiary details for qr code generation
